refactor(envelope): improve queries for spent sum
